New ChunkStore format.
Due to an off by one error that has existed in ChunkStore for the past two years,
a major exploit has been found in our system.  This commit creates alters our
handling of the ChunkStore to patch the exploit, however, it requires a full
conversion of existing ChunkStore into the new format.  Because of several factors
in the nature of the ChunkStore bug, the conversion cannot be done from within
the plugin, and instead has been packaged as a standalone program within the jar.

It is highly recommended that you convert your ChunkStore as soon as possible.
However, the old format will still function(Albiet with the exploit still possible)